# partB/openapi.yaml
openapi: 3.0.3
info:
  title: Mini Library API
  version: 1.0.0
paths:
  /api/health:
    get:
      summary: Health check
      responses:
        "200":
          description: Service is healthy
  /api/auth/login:
    post:
      summary: Login as an administrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login succeeded and an httpOnly JWT cookie was set
        "400":
          description: Missing credentials
        "401":
          description: Invalid credentials
  /api/auth/logout:
    post:
      summary: Logout current administrator
      responses:
        "200":
          description: Auth cookie cleared
  /api/auth/me:
    get:
      summary: Return the current administrator from the auth cookie
      responses:
        "200":
          description: Current administrator
        "401":
          description: Missing or invalid auth cookie
  /api/books:
    get:
      summary: List books
      parameters:
        - in: query
          name: query
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: availability
          schema:
            type: string
            enum: [all, available, unavailable]
      responses:
        "200":
          description: Book list
    post:
      summary: Create book
      responses:
        "201":
          description: Book created
        "400":
          description: Validation error
  /api/members:
    get:
      summary: List members
      responses:
        "200":
          description: Member list
    post:
      summary: Create member
      responses:
        "201":
          description: Member created
        "400":
          description: Validation error
  /api/loans:
    get:
      summary: List loans
      responses:
        "200":
          description: Loan list
    post:
      summary: Create loan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookId, memberId]
              properties:
                bookId:
                  type: integer
                memberId:
                  type: integer
                days:
                  type: integer
                  minimum: 1
                  maximum: 365
      responses:
        "201":
          description: Loan created
        "400":
          description: Loan blocked
  /api/loans/{id}/extend:
    patch:
      summary: Extend an active loan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                days:
                  type: integer
                  minimum: 1
                  maximum: 365
                  default: 7
      responses:
        "200":
          description: Loan extended
        "400":
          description: Invalid extension period
        "404":
          description: Active loan not found
  /api/loans/{id}/return:
    patch:
      summary: Return a loan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan returned
        "404":
          description: Active loan not found
  /api/dashboard/summary:
    get:
      summary: Dashboard summary
      responses:
        "200":
          description: Aggregated counts